% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{binaryPlot}
\alias{binaryPlot}
\title{binaryPlot}
\usage{
binaryPlot(x, axlabs = TRUE, ...)
}
\arguments{
\item{x}{matrix}

\item{axlabs}{Binary, should the axis labels be shown.}

\item{...}{title, xlabs, ylabs.}
}
\description{
Plot for binary matrices, especially adjacency matrices.
}
\details{
binaryPlot
}
