% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMetrics.R
\name{clustCoef}
\alias{clustCoef}
\title{clustCoef}
\usage{
clustCoef(x)
}
\arguments{
\item{x}{adjacency matrix}
}
\value{
scaler
}
\description{
Calculates the cluster coefficient from a network
}
\details{
Given a network in the form of adjacency matrix, this calculates the cluster
    coefficient. For a definition of cluster coefficient, please refer to
    the igraph documentation.
}
\examples{
clustCoef(beach[[1]][, ])
}
\author{
Abhirup
}
