% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnr.R
\docType{package}
\name{dnr}
\alias{dnr}
\alias{dnr-package}
\title{dnr: A package for simulating dynamic networks using ERGM family models.}
\description{
This package provides functions for fitting lagged exponential family models on dynamic network data, simulation from the models and model diagnostics. \cr
}
\keyword{internal}
