% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_bayes.R
\name{engine_bayes}
\alias{engine_bayes}
\title{engine_bayes}
\usage{
engine_bayes(start_network, inputcoeff, ns, model.terms, model.formula,
  graph_mode, group, intercept, exvar, maxlag, lagmat, ylag, lambda = NA,
  method = "bayesglm", alpha.glmnet, paramout = TRUE, Theta = NA)
}
\arguments{
\item{start_network}{Initial list of networks}

\item{inputcoeff}{coefficient vector}

\item{ns}{number of time points for simulation}

\item{model.terms}{model terms in formula}

\item{model.formula}{model formula (ergm)}

\item{graph_mode}{'digraph' by default}

\item{group}{group terms}

\item{intercept}{intercept terms}

\item{exvar}{extraneous covariates}

\item{maxlag}{maximum lag}

\item{lagmat}{lag matrix}

\item{ylag}{lag vector for network lag terms}

\item{lambda}{NA}

\item{method}{'bayesglm' by default}

\item{alpha.glmnet}{NA}

\item{paramout}{T/F parameter estimation is returned.}

\item{Theta}{= prior probability matrix.}
}
\description{
Implementation of simulation engine for dynamic networks using smoothing estimates of change statistics.
}

