% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateVertex.R
\name{engineVertex}
\alias{engineVertex}
\title{Simulation Engine for dynamic Vertex case.}
\usage{
engineVertex(InputNetwork, numSim, maxLag, VertexStatsvec = rep(1,
  nvertstats), VertexLag = rep(1, maxLag), VertexLagMatrix = matrix(1,
  maxLag, length(VertexStatsvec)), EdgeModelTerms, EdgeModelFormula,
  EdgeGroup = NA, EdgeIntercept = c("edges"), EdgeExvar = NA,
  EdgeLag = rep(1, maxLag), EdgeLagMatrix = matrix(1, maxLag,
  length(EdgeModelTerms)), regMethod = "bayesglm", paramout = TRUE)
}
\arguments{
\item{InputNetwork}{List of input networks}

\item{numSim}{number of time points to simulate}

\item{maxLag}{maximum Lag}

\item{VertexStatsvec}{Binary vector for vertex model.}

\item{VertexLag}{vector of lag for vertex}

\item{VertexLagMatrix}{matrix of lags for vertex stats.}

\item{EdgeModelTerms}{Edge Model terms}

\item{EdgeModelFormula}{Edge model formula}

\item{EdgeGroup}{edge group term}

\item{EdgeIntercept}{edge intercept}

\item{EdgeExvar}{edge extraneous variable}

\item{EdgeLag}{edge Lag vector}

\item{EdgeLagMatrix}{edge lag matrix}

\item{regMethod}{regression method. "bayesglm" by default}

\item{paramout}{T/F on if regression needs to run.}
}
\value{
List with following elements:
SimNetwork: Output Networks
EdgeParameterMat: Matrix of edge parameter
VertexParameterMat: Matrix of Vertex parameters.
}
\description{
Simulation engine for dynamic vertex case.
}
\examples{
## test function:
simResult <- engineVertex(InputNetwork = beach,
                          numSim = 5,
                          maxLag = 3,
                          EdgeModelTerms = c("triadcensus.003",
                                             "triadcensus.012",
                                             "triadcensus.102",
                                             "triadcensus.021D"),
                          EdgeModelFormula = net ~ triadcensus(0:3)
                          )
}
\author{
Abhirup
}

