% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMetrics.R
\name{expdeg}
\alias{expdeg}
\title{expdeg}
\usage{
expdeg(x)
}
\arguments{
\item{x}{adjacency matrix}
}
\value{
scaler
}
\description{
Calculate the expectation of degree distribution of network
}
\details{
Given a network in adjacency matrix form, this calculates the expected
    degree statsitic using igraph degree distribution function.
}
\examples{
expdeg(beach[[1]][, ])
}
\author{
Abhirup
}
