% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMetrics.R
\name{ntriangles}
\alias{ntriangles}
\title{ntriangles}
\usage{
ntriangles(x)
}
\arguments{
\item{x}{square matrix (adjacency matrix)}
}
\value{
scaler, number of triangles
}
\description{
Calculate number of triangles of a network
}
\details{
This function calculates the number of triangles in a network given an
    adjacency matrix. We use igraph for this.
}
\examples{
ntriangles(beach[[1]][, ])
}
\author{
Abhirup
}
