% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramest.R
\name{paramest}
\alias{paramest}
\title{Parameter estimation for static vertex case.}
\usage{
paramest(input_network, model.terms, model.formula, graph_mode = "digraph",
  group, intercept = c("edges"), exvar = NA, maxlag = 3,
  lagmat = matrix(sample(c(0, 1), (maxlag + 1) * length(model.terms), replace
  = T), ncol = cdim), ylag = rep(1, maxlag), lambda = NA,
  method = "glmnet", alpha.glmnet = 1, paramout = TRUE)
}
\arguments{
\item{input_network}{Input network.}

\item{model.terms}{model terms, must be ERGM terms expanded.}

\item{model.formula}{ERGM formula for each time point.}

\item{graph_mode}{'digraph' by default for bidirectional.}

\item{group}{grouping covariates for vertices.}

\item{intercept}{intercept terms.}

\item{exvar}{Extraneous variables}

\item{maxlag}{maximum lag.}

\item{lagmat}{Matrix of dimension (maxlag+1)x(length(model.terms))}

\item{ylag}{lag vectors of length=maxlag.}

\item{lambda}{NA}

\item{method}{Regression method, default is 'bayesglm'}

\item{alpha.glmnet}{if regularization is used. not needed for bayesglm.}

\item{paramout}{TRUE by default. if parameters are needed.}

\item{model.terms}{model terms vector.}

\item{exterm:}{string, the term to be excluded. For now I am using 'logsize'}

\item{model.formula}{model formula in ergm.}

\item{exterm:}{string, the term to be excluded. For now I am using 'logsize'}

\item{output.edge}{XY matrix for estimation.}

\item{method:}{one of 'glmnet', 'glm' or 'bayesglm'}
}
\value{
list with elements:
  coef: coefficients
  mplematfull: full martix of change statistics
  mplemat: subset of matrix of change statistics

model.terms.new

Exclude a term from the formula

term_formula.new
TODO: should support lagmatrix, and model.formula.new output?

list with elements:
       coef.edge: matrix of coeff vector for each edge
       lambda: lambda
       std.err
}
\description{
Parameter estimation for the static vertex case.
}
\examples{
input_network=rdNets[1:6]
model.terms=c("triadcensus.003", "triadcensus.012", "triadcensus.102", "triadcensus.021D", "gwesp");
model.formula = net~triadcensus(0:3)+gwesp(alpha=0, fixed=FALSE, cutoff=30)-1;
graph_mode='digraph';
group='dnc';
alpha.glmnet=1
directed=TRUE;
method <- 'bayesglm'
maxlag <- 3
lambda=NA
intercept = c("edges")
cdim <- length(model.terms)
lagmat <- matrix(sample(c(0,1),(maxlag+1)*cdim,replace = TRUE),ncol = cdim)
ylag <- rep(1,maxlag)
exvar <- NA
out <- paramest(input_network,model.terms, model.formula,
                graph_mode='digraph',group,intercept = c("edges"),exvar=NA,
                maxlag = 3,
                lagmat = matrix(sample(c(0,1),(maxlag+1)*cdim,
                                       replace = TRUE),ncol = cdim),
                ylag = rep(1,maxlag),
                lambda = NA, method='bayesglm',
                alpha.glmnet=1)
}
\author{
Abhirup
}

