% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramVertex.R
\name{paramVertex}
\alias{paramVertex}
\title{Parameter estimation for Vertex dynamics}
\usage{
paramVertex(InputNetwork, VertexStatsvec = rep(1, 8), maxLag,
  VertexLag = rep(1, maxLag), VertexLagMatrix = matrix(1, maxLag,
  length(VertexStatsvec)), EdgeModelTerms, EdgeModelFormula, EdgeGroup,
  EdgeIntercept = c("edges"), EdgeExvar = NA, EdgeLag = rep(1, maxLag),
  EdgeLagMatrix = matrix(1, maxLag, length(EdgeModelTerms)),
  regMethod = "bayesglm", paramout = FALSE)
}
\arguments{
\item{InputNetwork}{list of networks.}

\item{VertexStatsvec}{binary vector of size 8.}

\item{maxLag}{maximum lag, numeric.}

\item{VertexLag}{binary vector of length maxLag.}

\item{VertexLagMatrix}{binary matrix of size maxLag x 8.}

\item{EdgeModelTerms}{Model terms in edge model.}

\item{EdgeModelFormula}{Model formula in edge model.}

\item{EdgeGroup}{Group terms in edge model.}

\item{EdgeIntercept}{Intercept for edge model.}

\item{EdgeExvar}{Extraneous variable for edge model.}

\item{EdgeLag}{binary vector of length maxLag.}

\item{EdgeLagMatrix}{binary matrix of dim maxLag x length(EdgeModelTerms)}

\item{regMethod}{Regression method. default: "bayesglm"}

\item{paramout}{T/F Should the parameter estimates be returned?}
}
\value{
list with following elements:
EdgeCoef: edge coefficients.
Edgemplematfull: MPLE matrix from edges.
Edgemplemat: Subsetted MPLE matrix.
VertexCoef: Coefficients from vertex.
Vstats: Vertex statistics matrix.
}
\description{
Parameter estimation fro dynamic vertex case. The interface remaining almost identical to the static vertex one.
}
\examples{
out <- paramVertex(InputNetwork = beach,
     maxLag = 3,
     EdgeModelTerms = c("triadcensus.003", "triadcensus.012",
      "triadcensus.102","triadcensus.021D"),
     EdgeModelFormula = net ~ triadcensus(0:3),
     EdgeGroup = "group1",
     EdgeIntercept = c("edges"),
     paramout = TRUE)
}
\author{
Abhirup
}

