% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramVertexOnly.R
\name{paramVertexOnly}
\alias{paramVertexOnly}
\title{Parameter estimation for Vertex model only for a list of dynamic networks.}
\usage{
paramVertexOnly(InputNetwork, VertexStatsvec = rep(1, nvertexstats), maxLag,
  VertexLag = rep(1, maxLag), VertexLagMatrix = matrix(1, maxLag,
  length(VertexStatsvec)), regMethod = "bayesglm")
}
\arguments{
\item{InputNetwork}{Input network list.}

\item{VertexStatsvec}{Binary vector of size 9, indicating vertex model.}

\item{maxLag}{maximum lag.}

\item{VertexLag}{Binary vector of size maxLag, indicating Lag terms in the model.}

\item{VertexLagMatrix}{Binary matrix indicating lagged vertex statistics in
the model.}

\item{regMethod}{one of "glm", "glmnet", "bayesglm"}
}
\value{
List of 3 elements:\cr
VertexFit: Output from regEngine. \cr
VertexStats: Subsetted vertex stats matrix. \cr
VertexStatsFull: Full matrix of vertex stats.
}
\description{
Parameter estimation for Vertex model only for a list of dynamic networks.
}
\author{
Abhirup
}

