% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdNets}
\alias{rdNets}
\title{Blog citation network}
\format{A list with 484 elements. Each element is a network of size 47 number of vertices.}
\source{
Butts, C. T. and B. R. Cross (2009). Change and external events in computer-mediated citation networks, English language weblogs and the 2004 u.s. electoral cycle. The Journal of Social Structure 10 (3), 1-29.
}
\usage{
rdNets
}
\description{
A data set of temporal inter and intra group blog citation network, with fixed number of vertex.
}
\keyword{datasets}
