% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{regEngine}
\alias{regEngine}
\title{General purpose regression engine for the methods bayesglm, glm and glmnet}
\usage{
regEngine(XYdata, method = "bayesglm", regIntercept = FALSE, lambda = NA,
  alpha = 1)
}
\arguments{
\item{XYdata}{matrix with X and Y columns. First column is named as y, other columns are X.}

\item{method}{string among ("glm", "glmnet", "bayesglm").}

\item{regIntercept}{Logical. Should intercept be included in the model?}

\item{lambda}{for method "glmnet".}

\item{alpha}{for "glmnet"}
}
\value{
list with elements: coef, se, lambda, fit (Coefficients, SE, lambda, if used, fit object.)
}
\description{
General purpose regression engine for the methods bayesglm, glm and glmnet
}
\author{
Abhirup
}
