% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMetrics.R
\name{vdegree}
\alias{vdegree}
\title{vdegree}
\usage{
vdegree(x)
}
\arguments{
\item{x}{Adjacency matrix.}
}
\value{
vector of length number of vertices.
}
\description{
Calculates the degree of each vertices.
}
\details{
Given a network as adjacency matrix, calculate degree stats for each vertex.
}
\examples{
vdegree(beach[[1]][, ])
}
\author{
Abhirup
}
