% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnr.R
\docType{package}
\name{dnr}
\alias{dnr}
\title{dnr: A package for simulating dynamic networks using ERGM family models.}
\description{
This package provides functions for fitting lagged exponential family models on dynamic network data, simulation from the models and model diagnostics. \cr
}
\section{note}{

This package was developed with help from ARO YIP award \#W911NF-14-1-0577. \cr
}

\section{references}{

Abhirup Mallik & Zack W. Almquist (2019) Stable Multiple Time Step Simulation/Prediction From Lagged Dynamic Network Regression Models, Journal of Computational and Graphical Statistics, 28:4, 967-979, <DOI: 10.1080/10618600.2019.1594834>. \cr
Abhirup Mallik and Zack W. Almquist (2017). "An R Package for Dynamic Network Regression." Working paper. University of Minnesota. \cr
Zack W. Almquist and Carter T. Butts (forthcoming). "Dynamic Network Analysis with Missing Data: Theory and Methods." Statistica Sinica. <doi:10.5705/ss.202016.0108>. \cr
Zack W. Almquist and Carter T. Butts. (2013). "Dynamic Network Logistic Regression: A Logistic Choice Analysis of Inter- and Intra-group Blog Citation Dynamics in the 2004 US Presidential Election." Political Analysis, 21(4), 430-448. \cr
Zack W. Almquist and Carter T. Butts. (2014). "Bayesian Analysis of Dynamic Network Regression with Joint Edge/Vertex Dynamics." In Bayesian Inference in the Social Sciences. Ed. by I. Jeliazkov and X.-S. Yang. Hoboken, New Jersey: John Wiley & Sons. \cr
}

\keyword{internal}
